Atualização

+++ app/Http/Controllers/AlunoController.php
@@ -69,7 +69,7 @@ class AlunoController extends Controller
             $post_certificado->save();
 
             if($post_certificado){
-                return redirect('/aluno')->with('sucesso', 'O certificado foi enviado com sucesso!');
+                return redirect('/aluno')->with('sucesso', 'O certificado foi enviado com sucesso');
             }else{
                 return redirect('/aluno')->with('erro', 'Ocorreu um erro ao tentar enviar seu certificado. Por favor tente novamente');
             }
@@ -96,13 +96,20 @@ class AlunoController extends Controller
 
     }
 
-    /* Função para mostrar detalhes dos certificados para serem editados */
-    public function listaDados($id)
+    /* Função para mostrar detalhes dos certificados para serem editados 
+     * Envia o resultado da consulta para a função ajax "enviaDadosViewAluno" que esta no arquivo
+     * "public/js/lista_dados.js"
+    */
+    public function editarDetalhesCertificado($id)
     {
         $consulta = DB::table('certificados')->SELECT('id_certificado', 'tipo', 'inicio', 'termino', 'carga_horaria')->where('id_certificado', $id)->first();
         return response()->json($consulta);                    
     }
 
+    /* Função para mostrar detalhes dos certificados para serem excluidos 
+     * Envia o resultado da consulta para a função ajax "excluirCertificado" que esta no arquivo
+     * "public/js/lista_dados.js"
+    */
     public function ajaxCertificadoDelete($id)
     {
         $certificado_delete = DB::table('certificados')->SELECT('id_certificado', 'nome_certificado')->where('id_certificado', $id)->first();
@@ -110,7 +117,7 @@ class AlunoController extends Controller
     }
 
     /* Função para fazer o download do certificado */
-    public function download($id)
+    public function downloadCertificado($id)
     {
         //$id_user = Auth::user()->id;
         $arquivo = Certificados::where('user_id', 1)->where('id_certificado', $id)->first();
@@ -158,7 +165,7 @@ class AlunoController extends Controller
         $id_user_auth = 1; //Auth::user()->id;
         //dd($id_user_auth);
         $id_user_certificado = Certificados::all('user_id');
-        dd($id_user_certificado);
+        //dd($id_user_certificado);
 
         if($id_user_auth == $id_user_certificado){
 
@@ -182,7 +189,7 @@ class AlunoController extends Controller
         //dd($post);
         
         if($post){
-            return redirect('/aluno')->with('sucesso','As informações do certificado foram alteradas com sucesso!');
+            return redirect('/aluno')->with('sucesso','As informações do certificado foram alteradas com sucesso');
         }else{
             return redirect('/aluno')->with('erro','Não foi possível editar o certificado. Por favor tente novamente');
         }
@@ -202,11 +209,10 @@ class AlunoController extends Controller
 
         //$id_user = Auth::user()->id;
         $post = $request->id_certificado_excluir;
-        $post = Certificados::where('user_id', 1
-        )->where('id_certificado', $post)->delete();
+        $post = Certificados::where('user_id', 1)->where('id_certificado', $post)->delete();
         
         if($post){
-            return redirect('/aluno')->with('sucesso', 'O certificado foi excluido com sucesso!');
+            return redirect('/aluno')->with('sucesso', 'O certificado foi excluido com sucesso');
         }else{
             return redirect('/aluno')->with('erro', 'Ocorreu um erro ao tentar excluir o certificado. Por favor tente novamente');
         }
